requirements.txt
Flask-Migrate
faster_whisper==1.1.1
flasgger==0.9.7.1
Flask==3.1.1
flask_sqlalchemy==3.1.1
torch
torchvision
torchaudio
python-dotenv
pyjwt==2.10.1